test: 몽고 디비 접근은 메세지 컨트롤러 테스트에서만 하도록 수정
